Make base compile: define Sensorino::die, remove RX interrupt.
The UART RX interrupt is already registered by the Arduino UART
libraries, our implementation would conflict.

Sensorino::die is used by Message.cpp but not defined in the
Base since we don't compile Sensorino.cpp into the binary (if the
Sensorino and Base files are split into separate libraries which
is necessary anyway).