A dirty workaround to make Base compile out of the box.
Apparently when the Arduino IDE starts up it looks over all the
subdirectories in libraries/ and caches all .h files.  If one of the
files is #included from a sketch, the whole directory will be linked
with the sketch's binary.  With Sensorino.h both inside
libraries/Sensorino/ and
libraries/Base,
it would try to compile and link all files in both libraries, causing
duplicate symbol conflicts.