Make sure SS pin is an output at all times.
* In the nRF library make sure the SS pin is an output before setting
the MSTR bit.

* Change the default nRF24L01+ pin mapping so that CSN is on SS and
interrupt is on what CSN was previously (swap CSN and interrupt).  In
general the default pin mapping should only be there for illustration
purposes so users know where to configure their real nRF24L01+ wiring,
and we (the developers) should not commit our local configs to the
repository, but in this case the interrupt pin being same as the SS
pin was preventing the interrupt from being usable at all.

* Fix the comment about clock divider in spi_init().

* Don't include the Arduino SPI library since we're not using it
while we're using it for the moment, until we can switch back to real
RadioHead.