BER-compatible TLV multibyte tags.
Decoding and encoding support for types > 127, with an optional way to
skip the first byte which for us is always the same, at the cost of
losing BER compatibility (which we never had so far anyway, and probably
won't have for things like float encoding for example).

Another way to do this would be to use the Application specific tag
Class but use the 5 bits in the initial byte instead of setting them to
all-ones, which seems like it is allows provided we're not using the
Universal class.  This would give us 1-byte tag for the most common
types while retaining BER compatibility.