Mask unwanted bits in the result of ~.
GCC optimizes away ~*gpio used as a condition, possibly because
apparently the ~ operator ignores the operand type unlike other
operators.  It seems the operand is always cast to signed in and the
result is a signed int.  This was causing a very tricky bug to occur
that could only be seen in the resulting assembly.  If *gpio was
marked volatile, it would still only compile to "ld <reg>, gpio" with
optimisation.